Remove EntitiesMap, rename file to Entity
