+++ src/types/Entity.ts
@@ -1,0 +1,3 @@
+export default interface Entity<T extends number | string = number> {
+    id: T;
+}
