+++ src/types/index.ts
@@ -1,3 +1,3 @@
 export { default as Coverage } from './Coverage';
 export { default as Pagination } from './Pagination';
-export { Entity } from './EntitiesMap';
+export { default as Entity } from './Entity';
